Update AZ-204_lab_05.md
This is confusing to lab students ... if it is on one line it makes sense and will not return an error.  Flatten two lines into just one and it works fine. I believe you could end the first line with ^ which should work as well, but seems easiest to just make it one line.